@@ -6,90 +6,82 @@
   "author": "Martin",
   "main": "index.js",
   "scripts": {
-    "build": "react-scripts build",
-    "dump-box-data": "node -r esm ./scripts/dump-box-data.js",
-    "eject": "react-scripts eject",
+    "build": "node scripts/build.js",
+    "dump:box-data": "ts-node ./scripts/dump-box-data.ts",
     "lint:style": "stylelint 'src/**/*.{css,js,jsx}'; exit 0",
-    "lint:js": "eslint 'libs/**/*.{js,jsx}'",
-    "lint:js:fix": "npm run lint:js -- --fix",
     "lint:ts": "eslint 'src/**/*.{ts,tsx}'",
     "lint:ts:fix": "npm run lint:ts -- --fix",
     "format": "prettier --write './**/*.{ts,tsx,js,jsx,json}'",
     "format-and-lint": "npm run format && npm run lint:style && npm run lint:ts:fix",
-    "start": "BROWSER=none REACT_APP_MODE=live react-scripts start",
-    "start:demo": "BROWSER=none REACT_APP_MODE=dev react-scripts start",
-    "test": "react-scripts test",
-    "test:lib": "jest './libs/*' --colors --coverage --watchAll"
+    "start": "VITE_APP_MODE=prod vite",
+    "start:demo": "VITE_APP_MODE=dev vite",
+    "test": "node scripts/test.js",
+    "test:schema": "vitest './src/schema/boxFields/box.schema.ts'"
   },
   "dependencies": {
-    "@babel/core": "^7.9.6",
-    "@babel/plugin-transform-typescript": "^7.9.6",
-    "@babel/preset-env": "^7.9.6",
-    "@babel/preset-react": "^7.9.4",
-    "axios": "^0.19.2",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^24.9.0",
-    "babel-plugin-macros": "^2.8.0",
-    "babel-plugin-transform-export-extensions": "^6.22.0",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react-app": "^9.1.2",
-    "date-fns": "^2.13.0",
-    "dotenv": "^8.2.0",
-    "esm": "^3.2.25",
-    "fetch-mock": "^9.5.1",
-    "fs-extra": "^9.0.0",
-    "libnpx": "^10.2.3",
-    "lodash": "^4.17.15",
-    "moxios": "^0.4.0",
-    "node-fetch": "^2.6.0",
-    "prettier": "^2.0.5",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-scripts": "^3.4.1",
-    "react-transition-group": "^1.0",
-    "regenerator-runtime": "^0.13.5",
-    "styled-components": "^5.1.0",
-    "styled-normalize": "^8.0.7",
-    "typescript": "^3.8.3",
-    "uid": "^1.0.0"
+    "@tanstack/react-query": "^5.25.0",
+    "@tanstack/react-query-devtools": "^5.25.0",
+    "@typescript-eslint/parser": "^7.1.1",
+    "@vitejs/plugin-react": "^4.2.1",
+    "@zodios/core": "^10.9.6",
+    "browserslist": "^4.23.0",
+    "date-fns": "^3.3.1",
+    "dotenv": "^16.4.5",
+    "eslint": "^8.57.0",
+    "fetch-mock": "^9.11.0",
+    "fs-extra": "^11.2.0",
+    "jest": "^29.7.0",
+    "jest-resolve": "^29.7.0",
+    "lodash": "^4.17.21",
+    "modern-normalize": "^2.0.0",
+    "msw": "^2.2.3",
+    "node-fetch": "^3.3.2",
+    "postcss": "^8.4.35",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-loader": "^8.1.1",
+    "postcss-normalize": "^10.0.1",
+    "postcss-preset-env": "^9.5.0",
+    "prettier": "^3.2.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-transition-group": "1.x",
+    "styled-components": "^6.1.8",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.2",
+    "uid": "^2.0.2",
+    "vite": "^5.1.5",
+    "vite-plugin-svgr": "^4.2.0",
+    "vite-tsconfig-paths": "^4.3.1",
+    "vitest": "^1.3.1",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/jest": "^25.2.1",
-    "@types/lodash": "^4.14.150",
-    "@types/node": "^13.13.5",
-    "@types/prop-types": "^15.7.3",
-    "@types/react": "^16.9.35",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-transition-group": "^4.2.4",
-    "@types/styled-components": "^5.1.0",
-    "@typescript-eslint/eslint-plugin": "^2.33.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "enzyme-to-json": "^3.4.4",
-    "eslint": "6.8.0",
-    "eslint-config-airbnb": "18.1.0",
-    "eslint-config-react-app": "^5.2.1",
-    "eslint-plugin-flowtype": "4.7.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-lodash": "^7.1.0",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "2.5.0",
-    "jest": "^24.9.0",
-    "jest-styled-components": "^7.0.2",
-    "ndb": "^1.1.5",
-    "react-test-renderer": "^16.13.1",
-    "sinon": "^9.0.2",
-    "sort-package-json": "^1.42.2",
-    "stylelint": "^13.3.3",
+    "@types/jest": "^29.5.12",
+    "@types/lodash": "^4.14.202",
+    "@types/node": "^20.11.25",
+    "@types/react": "^18.2.64",
+    "@types/react-dom": "^18.2.21",
+    "@types/react-transition-group": "^4.4.10",
+    "@types/styled-components": "^5.1.34",
+    "@typescript-eslint/eslint-plugin": "^7.1.1",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-lodash": "^7.4.0",
+    "eslint-plugin-react": "^7.34.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "jest-styled-components": "^7.2.0",
+    "stylelint": "^16.2.1",
     "stylelint-config-rational-order": "^0.1.2",
-    "stylelint-config-standard": "^20.0.0",
+    "stylelint-config-standard": "^36.0.0",
     "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-order": "^4.0.0",
+    "stylelint-order": "^6.0.4",
     "stylelint-processor-styled-components": "^1.10.0"
   },
-  "proxy": "http://fritz.box/"
+  "msw": {
+    "workerDirectory": [
+      "public"
+    ]
+  }
 }
