@@ -1,7 +1,6 @@
 import React, {
-  useState, useRef, useEffect, createRef,
+  useState, useRef, useEffect, createRef, useContext,
 } from 'react';
-import PropTypes from 'prop-types';
 
 // import { throttle } from 'lodash';
 import { NavBarEntry } from '../NavBarEntry';
@@ -9,28 +8,27 @@ import { BoxDataContext } from '../DataProvider';
 
 import * as Styles from './NavBar.styles';
 
-const NavBar: React.FC<{}> = (): JSX.Element => {
-  const state = React.useContext(BoxDataContext);
+const NavBar = () => {
+  const state = useContext(BoxDataContext);
 
   const [offset, setOffset] = useState(0);
-  const [width, setWidth] = useState('auto'); // prevent css transition on load
-
+  const [width, setWidth] = useState<string | number>('auto'); // prevent css transition on load
   const oldIndex = useRef(-1);
   const showIndicator = useRef(true);
-  const activeElement = createRef();
+  const activeElement = createRef<HTMLLIElement>();
 
+  const {
+    visibleComponents,
+    currentIndex,
+    isUpdating,
+    boxData,
+  } = state;
   const height = 5;
-
-  const componentsToShow = 'componentsToShow' in state ? state.componentsToShow : [];
-  const currentIndex = 'currentIndex' in state ? state.currentIndex : 0;
-
-  const isUpdating = 'isUpdating' in state ? state.isUpdating : true;
 
   function handleNavigation(newCurrentIndex: number): void {
     oldIndex.current = currentIndex;
 
     state.updateCurrentIndex(newCurrentIndex);
-    // setCurrentIndex(index);
   }
 
   function updateIndicator(): void {
@@ -57,36 +55,29 @@ const NavBar: React.FC<{}> = (): JSX.Element => {
     }
   });
 
+  if (!boxData) {
+    return null;
+  }
+
   return (
-    <Styles.NavBar reservedSpaceTop={height} isUpdating={isUpdating}>
+    <Styles.NavBar $reservedSpaceTop={height} $isUpdating={isUpdating}>
       {showIndicator.current && <Styles.Indicator offset={offset} width={width} height={height} />}
-      <Styles.NavBarList isRow={showIndicator}>
-        {componentsToShow.map((entry, index) => (
+      <Styles.NavBarList $isRow={showIndicator.current}>
+        {visibleComponents.map((entry, index) => (
           <NavBarEntry
             index={index}
             entry={entry}
-            isActive={currentIndex === index}
-            handleNavigation={handleNavigation}
+            $isActive={currentIndex === index}
+            handleNavigation={(newCurrentIndex: number): void => handleNavigation(newCurrentIndex)}
             // only active element gets ref otherwise last child always active
             activeElementRef={currentIndex === index ? activeElement : null}
-            key={index}
-            isFullWidth={!showIndicator}
+            key={entry}
+            $isFullWidth={!showIndicator}
           />
         ))}
       </Styles.NavBarList>
     </Styles.NavBar>
   );
-};
-
-const {
-  string, number, func, bool,
-} = PropTypes;
-
-NavBar.propTypes = {
-  componentsToShow: PropTypes.arrayOf(string).isRequired,
-  handleNavigation: func.isRequired,
-  currentIndex: number.isRequired,
-  isUpdating: bool.isRequired,
 };
 
 export { NavBar };
