@@ -1,10 +1,14 @@
-export interface StringMap {
-  [key: string]: string;
+export interface NavBarStyleProps {
+  $reservedSpaceTop: number;
+  $isUpdating: boolean;
 }
 
-export interface NavBarProps {
-  componentsToShow: any[];
-  currentIndex: number;
-  handleNavigation: any;
-  isUpdating: boolean;
+export interface IndicatorStyleProps {
+  offset: number;
+  width: string | number;
+  height: number;
+}
+
+export interface NavBarListStyleProps {
+  $isRow: boolean;
 }
