@@ -1,15 +1,15 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { upperFirst } from 'lodash';
 
 import * as Styles from './Slide.styles';
-import * as Types from './Slide.types';
+import { SlideProps } from './Slide.types';
 
-const Slide: React.FC<Types.SlideProps> = ({ title, text }): JSX.Element => (
+const Slide = ({ title, text }: SlideProps) => (
   <Styles.Wrapper>
     <Styles.TitleWrapper>{upperFirst(title)}</Styles.TitleWrapper>
-    <Styles.TextWrapper characterCount={text.length}>
+    <Styles.TextWrapper $characterCount={text.length}>
       {text.split('').map((character, index) => (
+        // eslint-disable-next-line react/no-array-index-key
         <span className="character" key={index}>
           {index === 0 ? upperFirst(character) : character}
         </span>
@@ -17,12 +17,5 @@ const Slide: React.FC<Types.SlideProps> = ({ title, text }): JSX.Element => (
     </Styles.TextWrapper>
   </Styles.Wrapper>
 );
-
-const { string } = PropTypes;
-
-Slide.propTypes = {
-  text: string.isRequired,
-  title: string.isRequired,
-};
 
 export { Slide };
