@@ -1,8 +1,8 @@
-export interface StringMap {
-  [key: string]: string;
-}
-
 export interface SlideProps {
   title: string;
   text: string;
+}
+
+export interface TextWrapperStyleProps {
+  $characterCount: number;
 }
