@@ -1,21 +1,14 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { ThemeProvider } from 'styled-components/macro';
-
-// import * as Types from './Theme.types';
+import React, { PropsWithChildren } from 'react';
+import { ThemeProvider } from 'styled-components';
 
 export const theme = {
   primaryColor: '#F2F2F2',
   secondaryColor: '#BDBDBD',
   tertiaryColor: '#080808',
-};
+} as const;
 
-export const Theme: React.FC<{}> = ({ children }): JSX.Element => (
+export type ThemeType = keyof typeof theme;
+
+export const Theme = ({ children }: PropsWithChildren) => (
   <ThemeProvider theme={theme}>{children}</ThemeProvider>
 );
-
-const { node } = PropTypes;
-
-Theme.propTypes = {
-  children: node.isRequired,
-};
