@@ -1,35 +1,34 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { useContext } from 'react';
 
 import { BoxDataContext } from '../DataProvider';
 
 import * as Styles from './UpdateBar.styles';
 
-const UpdateBar: React.FC<{}> = (): JSX.Element => {
-  const state = React.useContext(BoxDataContext);
+const UpdateBar = () => {
+  const state = useContext(BoxDataContext);
 
-  const isUpdating = 'isUpdating' in state ? state.isUpdating : true;
-  const isValid = 'isValid' in state ? state.isValid : false;
+  const {
+    isUpdating,
+    isValid,
+  } = state;
+
+  const shouldShowUpdateBar = isUpdating || !isValid;
 
   return (
-    <Styles.UpdateBar isUpdating={isUpdating}>
+    <Styles.UpdateBar $shouldShowUpdateBar={shouldShowUpdateBar}>
       {isUpdating ? (
         <Styles.Text>Updating!</Styles.Text>
       ) : (
         <Styles.Text>
-          {!isValid && <>Error!</>}
-          {!!isValid && <>Data loaded!</>}
+          {isValid ? (
+            <>Data loaded!</>
+          ) : (
+            <>Error!</>
+          )}
         </Styles.Text>
       )}
     </Styles.UpdateBar>
   );
-};
-
-const { bool } = PropTypes;
-
-UpdateBar.propTypes = {
-  isUpdating: bool.isRequired,
-  isValid: bool.isRequired,
 };
 
 export { UpdateBar };
