@@ -1,8 +1,3 @@
-export interface StringMap {
-  [key: string]: string;
-}
-
-export interface UpdateBarProps {
-  isUpdating: boolean;
-  isValid: boolean;
+export interface UpdateBarStyleProps {
+  $shouldShowUpdateBar: boolean;
 }
